Move SA tests to test_systolic_array.py
